Fully working WM model Needs to be changed for non-WM model

+++ centralDogma_surface_2022/random.h
@@ -9,7 +9,8 @@ inline std::random_device rd;
 inline std::seed_seq ss{rd(), rd(), rd(), rd(), rd(), rd(), rd(), rd()};
 inline std::mt19937 twister{ss};
 inline std::uniform_int_distribution typeInitializer{1, 3};
-inline std::uniform_int_distribution randomRowOrCol{1, Para::sys_nrow};
+inline std::uniform_int_distribution randomRowOrCol{
+    1, static_cast<int>(Para::sys_nrow)};
 
 // Index should be from 1 to nrow/ncol!!! Boundaries are on
 // index=0 and index=nrow+1/ncol+1
